refactor: extract executor thread into shared function_executor module
Move the global executor thread machinery out of scalar_function.c
into a new function_executor.c/h module so that it can be reused by
the upcoming AggregateFunction implementation.

- Add function_executor.{c,h} with rbduckdb_function_executor_ensure_started()
  and rbduckdb_function_executor_dispatch()
- Update scalar_function.c to call the shared dispatcher
- Include function_executor.h in ruby-duckdb.h before scalar_function.h
